Fix avatar profile if the receiptCard set to front
Problem:
Avatar profile is not touchable when the receiptCard
is set to parent or front.

Solution:
Implement a logic to check if the headerZIndex is set to 0 by default
if it is then I create another profile inside the receiptCard so that
this will be included also in the parent styles.